Raspberry/requirements-1.txt
picamera2
numpy>=1.21.2
dlib==19.22.1
firebase-admin==5.3.0
Pillow==8.4.0
opencv-python==4.5.5.64
RPi.GPIO>=0.7.0
pyttsx3>=2.90
gTTS>=2.2.4